+++ integrations/azure-devops/main.py
@@ -217,8 +217,9 @@ async def resync_repository_policies(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
 async def resync_workitems(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     azure_devops_client = AzureDevopsClient.create_from_ocean_config()
     config = cast(AzureDevopsWorkItemResourceConfig, event.resource_config)
+    wiql_filter = config.selector.build_wiql_filter()
     async for work_items in azure_devops_client.generate_work_items(
-        wiql=config.selector.wiql, expand=config.selector.expand
+        wiql=wiql_filter, expand=config.selector.expand
     ):
         logger.info(f"Resyncing {len(work_items)} work items")
         yield work_items